Moving java script fragments to files
