+++ koshmar/src/main/resources/javaScript/htmlElement/getAttributes.js
@@ -1,0 +1,3 @@
+function fun(element) {
+    return element.attributes;
+}
