+++ koshmar/src/main/resources/javaScript/htmlElement/getComputedStyle.js
@@ -1,8 +1,10 @@
-var hash = {};
-styleArgs = window.getComputedStyle(arguments[0], null);
-var len = styleArgs.length;
-for (var i = 0; i < len; i++) {
-    var style = styleArgs[i];
-    hash[style] = styleArgs.getPropertyValue(style);
+function fun(element) {
+    var hash = {};
+    var styleArgs = window.getComputedStyle(element, null);
+    var len = styleArgs.length;
+    for (var i = 0; i < len; i++) {
+        var style = styleArgs[i];
+        hash[style] = styleArgs.getPropertyValue(style);
+    }
+    return hash;
 }
-return hash;
